Line Cutter Data Log + Flight Bug Fixes

+++ system/tasks/cli/task_cli_sense.cpp
@@ -109,8 +109,8 @@ void cli_tasks::cliSense() {
   }
 #endif  // HAS_DEV(VBAT)
 #if HAS_DEV(PYRO_CONT)
-  snprintf(cliStr, sizeof(cliStr), "Pyro Continuity 1-%" PRIu8 ":",
-           (uint8_t)NUM_PYRO_CONT);
+  snprintf(cliStr, sizeof(cliStr), "Pyro Continuity 1-%" PRIu32 ":",
+           (uint32_t)(NUM_PYRO_CONT));
   cliSend(cliStr);
   for (int i = 0; i < NUM_PYRO_CONT; i++) {
     snprintf(cliStr, sizeof(cliStr), " %i", data->pyroContData[i]);
